+++ cosmosid/helpers/argument_validators.py
@@ -7,3 +7,9 @@ def uuid(param):
         return str(UUID(param))
     except Exception:
         raise ArgumentTypeError("Not a valid UUID!")
+
+
+def is_primer(value):
+    if isinstance(value, str) and not value.isalpha():
+        raise ArgumentTypeError("Only letters are allowed")
+    return value.upper() if isinstance(value, str) else value